pyproject.toml
[project]
name = "jdruwa-api"
version = "0.1.0"
description = "API GraphQL avec FastAPI et PostgreSQL"
authors = [
    {name = "Remi Petit",email = "remi.petit93370@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[all] (>=0.119.0,<0.120.0)",
    "strawberry-graphql (>=0.283.3,<0.284.0)",
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)"
]
